Consolidate sensors
